Fix rotor stepping; Implement double-step of midrotor
The simulator was failing to encode messages consistently with reliably
correct simulators see:
https://github.com/NationalSecurityAgency/enigma-simulator/issues/5

Identified root error being that the rotor was being stepped from the left rotor
rightwards on each keystroke, this is backwards. The Enigma machine steps the
first rotor by the entrywheel which is the rightmost rotor from perspective of
the operator. The simulator had the wirings reversed accordingly and is now
producing correct encodings.

Additionally added lacking double-step functionality for the mid-rotor.
The middle rotor will advance on the next step of the first rotor a
second time if the middle rotor notch is aligned with the pawl.